Restore anonymous implementation of GutenbergBridgeJS2Parent
